FIX UNWORKING DOUBLE FRONT NOTIFICATION:
Now the  client now he is already connected with the same account.
